refactor : 배틀 페이지 사이드바 컴포넌트 구조 개선
-  사이드바 컴포넌트 파일에서 헤더, 정보 탭을 가독성, 책임 분리를 위해 컴포넌트로 분리했습니다
- SidebarHedaer - 탭 전환, 닫기 버튼
- BattleInfoSection - 배틀 제목, 설명, 언어, 카테고리 표시 탭
